+++ BloodSeeker/BloodSeeker.csproj
@@ -170,6 +170,12 @@
     <Compile Include="Components\Calendar\CalendarButton.Designer.cs">
       <DependentUpon>CalendarButton.cs</DependentUpon>
     </Compile>
+    <Compile Include="Components\Calendar\CalendarFrame.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="Components\Calendar\CalendarFrame.Designer.cs">
+      <DependentUpon>CalendarFrame.cs</DependentUpon>
+    </Compile>
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -214,6 +220,9 @@
     </EmbeddedResource>
     <EmbeddedResource Include="Components\Calendar\CalendarButton.resx">
       <DependentUpon>CalendarButton.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Components\Calendar\CalendarFrame.resx">
+      <DependentUpon>CalendarFrame.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Form1.resx">
       <DependentUpon>Form1.cs</DependentUpon>
